requirements.txt
Django==5.0.14
djangorestframework==3.14.0
django-cors-headers==4.3.1
mssql-django==1.4